pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "faiss-gpu>=1.7.2",
    "google-genai>=1.22.0",
    "gradio>=5.34.2",
    "markitdown>=0.0.2",
    "nbconvert>=7.16.6",
    "nbformat>=5.10.4",
    "numpy<2",
    "ollama>=0.5.1",
    "opencv-python==4.9.0.80",
    "pymupdf4llm>=0.0.25",
    "trafilatura>=2.0.0",
]
